examples/rpc_response_streaming/pyproject.toml
[project]
name = "rpc-response-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.75.0",
    "grpcio-tools>=1.75.0",
]
